global variables working
