+++ src/renderer/components/Links.tsx
@@ -11,17 +11,19 @@
 import React, { useEffect } from 'react';
 import * as d3 from 'd3';
 // IMPORT HELPER FUNCTIONS
-import { Link, Services, Options, Simulation, SNode } from '../App.d';
+import { Link, Services, Options, SNode } from '../App.d';
 // IMPORT TYPES
 
 type Props = {
-  links: Link[];
   services: Services;
   options: Options;
-  simulation: Simulation;
 };
 
-const Links: React.FC<Props> = ({ links, services, options, simulation }) => {
+const Links: React.FC<Props> = ({ services, options }) => {
+  const {
+    simulation,
+    serviceGraph: { links },
+  } = window;
   useEffect(() => {
     simulation.force(
       'link',
