+++ src/renderer/components/Nodes.tsx
@@ -17,31 +17,19 @@ import {
 } from '../helpers/getSimulationDimensions';
 import { getStatic } from '../helpers/static';
 // IMPORT TYPES
-import {
-  SNode,
-  SetSelectedContainer,
-  Services,
-  Options,
-  Simulation,
-} from '../App.d';
+import { SNode, SetSelectedContainer, Services, Options } from '../App.d';
 // IMPORT COMPONENTS
 import NodePorts from './NodePorts';
 import NodeVolumes from './NodeVolumes';
 
 type Props = {
   services: Services;
-  nodes: SNode[];
   setSelectedContainer: SetSelectedContainer;
-  simulation: Simulation;
-  treeDepth: number;
   options: Options;
 };
 
 const Nodes: React.FC<Props> = ({
-  nodes,
   setSelectedContainer,
-  simulation,
-  treeDepth,
   services,
   options,
 }) => {
@@ -57,7 +45,7 @@ const Nodes: React.FC<Props> = ({
 
     //sets 'clicked' nodes back to unfixed position
     const dblClick = (d: SNode) => {
-      simulation.alphaTarget(0);
+      window.simulation.alphaTarget(0);
       d.fx = null;
       d.fy = null;
     };
@@ -65,7 +53,7 @@ const Nodes: React.FC<Props> = ({
     let drag = d3
       .drag<SVGGElement, SNode>()
       .on('start', function dragstarted(d: SNode) {
-        if (!d3.event.active) simulation.alphaTarget(0.3).restart();
+        if (!d3.event.active) window.simulation.alphaTarget(0.3).restart();
         d.fx = d3.event.x;
         d3.event.y;
       })
@@ -75,7 +63,7 @@ const Nodes: React.FC<Props> = ({
         d.fy = d3.event.y;
       })
       .on('end', function dragended(d: SNode) {
-        if (!d3.event.active) simulation.alphaTarget(0);
+        if (!d3.event.active) window.simulation.alphaTarget(0);
         d.fx = d.x;
         d.fy = d.y;
       });
@@ -84,7 +72,7 @@ const Nodes: React.FC<Props> = ({
     const nodeContainers = d3
       .select('.nodes')
       .selectAll('g')
-      .data<SNode>(nodes)
+      .data<SNode>(window.serviceGraph.nodes)
       .enter()
       .append('g')
       .on('click', (node: SNode) => {
@@ -97,7 +85,7 @@ const Nodes: React.FC<Props> = ({
         return (d.fx = getHorizontalPosition(d, width));
       })
       .attr('fy', (d: SNode) => {
-        return (d.fy = getVerticalPosition(d, treeDepth, height));
+        return (d.fy = getVerticalPosition(d, window.treeDepth, height));
       });
 
     // add names of services
