+++ src/renderer/index.tsx
@@ -12,7 +12,7 @@
 import * as React from 'react';
 import { render } from 'react-dom';
 import App from './App';
-import { Simulation } from '../renderer/App.d';
+import { Simulation, SGraph } from '../renderer/App.d';
 
 if (module.hot) {
   module.hot.accept();
@@ -21,6 +21,8 @@ if (module.hot) {
 declare global {
   interface Window {
     simulation: Simulation;
+    treeDepth: number;
+    serviceGraph: SGraph;
   }
 }
